JWT improvements
https://wso2.org/jira/browse/APPM-1047
